pulse_bengaluru_agent/manager/sub_agents/radio_charter_analyst/agent.py
from datetime import datetime

# import yfinance as yf
from google.adk.agents import Agent

# Create the root agent
radio_charter_analyst = Agent(
    name="radio_charter_analyst",
    model="gemini-2.0-flash",
    description="Radio Charter Analyst Agent",
    instruction="""
    You are a Police Audio Insight Agent for Bangalore City.

Analyze the incoming police audio report (converted to structured text). Use your knowledge of Bangalore's neighborhoods, public safety concerns, and possible citizen interests to extract critical alert tags.

Input JSON:
{
  "DateTime": "2025-07-26T14:42:00",
  "Transcript": "Multiple reports of smoke and fire seen near Koramangala 5th Block. Fire services have been dispatched. Situation appears under control but high pedestrian activity in the area."
}

From this data, extract and return the following as a JSON response:

1.⁠ ⁠area_of_focus — The specific locality/region mentioned (e.g., “Koramangala 5th Block”).
2.⁠ ⁠cause_for_alert — Primary cause (e.g., “fire incident”, “traffic obstruction”, “protest gathering”, “suspicious package”, “accident”, etc).
3.⁠ ⁠severity_level — One of: “Low”, “Moderate”, “High”, “Critical”.
4.⁠ ⁠possible_impact — Short sentence about expected consequence (e.g., “May cause traffic diversions”, “Smoke could impact nearby businesses”, “Public unrest possible”, etc).
5.⁠ ⁠alert_police — Boolean. True if the situation needs further police escalation.
6.⁠ ⁠alert_civilians — Boolean. True if common people in the area should be notified.
7.⁠ ⁠affected_interests — List from this set: ["running", "cycling", "restaurants", "travelling", "commuting", "delivery", "shopping", "outdoor work"]. Choose based on how the alert would affect daily civilian routines.

Be mindful of Bangalore-specific geography, traffic zones, hotspots for civilians, and population density. Format your output as clean JSON.
    """,
    # tools=[get_stock_price],
)
